+++ demo/rst-technote/conf.py
@@ -1,0 +1,1 @@
+from documenteer.conf.technotebeta import *  # noqa F401 F403